changing duplicate hit handling
Around 270,000 of the 2.2M-hit dataset (C23's
97 galaxies at L band) appear to be duplicates,
which should not affect the turboSETI analysis of
that paper but likely does affect HDBSCAN.

This update investigates the source of the
duplicates and concludes that they should be
removed for clustering. The analysis from the
final week of the REU will be repeated using the
slightly smaller dataset, before proceeding to
subsequent HDBSCAN iterations.