tSNE plus attempts at clustering non-anom hits
Non-anomalous hits are well clustered in each
batch in the batch_hdbscan_iterative.ipynb
process, but it remains a challenge to match
clusters to each other across batches.

The current aim is to connect each hit to its
cluster's centroid and then group the centroids,
on the assumption that matched clusters should
have similar centroids.

For reasons of speed and simplicity, I'm currently
doing this with k-means clustering. Obviously
there are issues with this, but hopefully it will
be good as a confirmative heuristic. HDBSCAN /
DBSCAN is unlikely to work due to the size of the
non-anomalous centroid data set (similar reasons
why we used batching in the first place).